packages/training/pyproject.toml
[project]
name = "baynext-ml-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Léo Grosjean", email = "leo.grosjean@live.fr" }
]
requires-python = ">=3.12"
dependencies = [
    "google-meridian>=1.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/baynext"]
